# Conflicts:
#	TASK.md
#	src/main/java/ua/duikt/learning/java/pro/spring/service/IssueService.java
#	src/main/java/ua/duikt/learning/java/pro/spring/service/impl/DetailsServiceImpl.java
#	src/main/java/ua/duikt/learning/java/pro/spring/service/impl/IssueServiceImpl.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint01/service/DetailsServiceTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/AttachmentControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/CommentControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/LabelControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/StatusControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/service/IssueServiceTest.java

+++ src/main/java/ua/duikt/learning/java/pro/spring/service/RoleService.java
@@ -9,7 +9,7 @@ import java.util.List;
  * email mykyta.sirobaba@gmail.com
  */
 public interface RoleService {
-    Integer createRole(String name);
+    Long createRole(String name);
 
     List<Role> getRoles();
 }
